Create a BaseCommandMixin
The way BaseCommand was constructed was a bit confusing, having two base.py files near each other.
It was also being used as a mixin, rather than a base class.

This doesn't change any functionality but moves the higher level BaseCommand to a mixin class which
makes thing easier to read and is more inline with how the class was being used in the first place.
